refactor: Overhaul with cleaner implementation
This removes the use of two-dimensional coordinates in favour of just
plain letter names (as referenced in graph.jpeg). This simplification
has been combined with the extraction of the puzzle data from the code
itself, leaving a much leaner and more straightforward way to solve the
problem.

Almost every type, function, class, and module has been traumatically
changed (for the better). This change was done slowly, with a pretty
extensive test suite informing how the refactor happened.